# Conflicts:
#	src/main/resources/mappers/assessment/IBIMS003B.xml

+++ src/main/java/com/nanuri/rams/business/common/dto/IBIMS110BDTO.java
@@ -10,10 +10,11 @@ import lombok.Setter;
 @Setter
 /* 딜심사책임준공내역 table(IBIMS110BDTO) DTO */
 public class IBIMS110BDTO {
+    
     private String     dealNo;                        // 딜번호
     private String     mtrDcd;                        // 안건구분코드
     private String     jdgmDcd;                       // 심사구분코드
-    private int        sn;                            // 일련번호
+    private BigDecimal sn;                            // 일련번호
     private String     rspsbCmplOgnDcd;               // 책임준공기관구분코드
     private String     sccoCrno;                      // 증권사법인등록번호
     private String     rnmCnfmNo;                     // 실명확인번호
@@ -27,4 +28,5 @@ public class IBIMS110BDTO {
     private String     hndTmnlNo;                     // 조작단말기번호
     private String     hndTrId;                       // 조작거래id
     private String     guid;                          // guid
+
 }
