+++ src/main/java/com/nanuri/rams/business/common/dto/IBIMS112BDTO.java
@@ -12,10 +12,11 @@ import lombok.Setter;
  위원회안건내역(협의체안건정보) Table.IBIMS112B DTO
 */
 public class IBIMS112BDTO {
+
     private String         cnsbDcd;                                         // 협의체구분코드
-    private int            cnsbSq;                                          // 협의체순번(회차)
+    private BigDecimal     cnsbSq;                                          // 협의체순번
     private String         rsltnYr;                                         // 결의년도
-    private int            sn;                                              // 일련번호(순서정보)
+    private BigDecimal     sn;                                              // 일련번호(순서정보)
     private String         dealNo;                                          // 딜번호
     private String         mtrDcd;                                          // 안건구분코드
     private String         jdgmDcd;                                         // 심사구분코드
@@ -33,4 +34,5 @@ public class IBIMS112BDTO {
     private String         hndTmnlNo;                                       // 조작단말기번호
     private String         hndTrId;                                         // 조작거래ID
     private String         guid;                                            // GUID
+
 }
